CCMinerMTP and CCMinerTpruvot create external dlls
