+++ NiceHashMiner.sln
@@ -110,6 +110,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "runnhmasadmin", "src\Tools\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RigIDPrinter", "src\Tools\RigIDPrinter\RigIDPrinter.csproj", "{138E97ED-8C5B-4C31-90F8-4260714A56E7}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CCMinerMTP", "src\Miners\CCMinerMTP\CCMinerMTP.csproj", "{C6E4069B-2305-4CDA-9F4E-0B79A8599935}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CCMinerTpruvot", "src\Miners\CCMinerTpruvot\CCMinerTpruvot.csproj", "{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -858,6 +862,38 @@ Global
 		{138E97ED-8C5B-4C31-90F8-4260714A56E7}.Release|Win32.Build.0 = Release|Any CPU
 		{138E97ED-8C5B-4C31-90F8-4260714A56E7}.Release|x64.ActiveCfg = Release|Any CPU
 		{138E97ED-8C5B-4C31-90F8-4260714A56E7}.Release|x64.Build.0 = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|Win32.Build.0 = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Debug|x64.Build.0 = Debug|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|Win32.ActiveCfg = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|Win32.Build.0 = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|x64.ActiveCfg = Release|Any CPU
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935}.Release|x64.Build.0 = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|Win32.Build.0 = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Debug|x64.Build.0 = Debug|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|Win32.ActiveCfg = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|Win32.Build.0 = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|x64.ActiveCfg = Release|Any CPU
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -904,6 +940,8 @@ Global
 		{DFC88036-796A-4046-B3B6-DE9FA56EDF28} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
 		{3712894A-0836-4CE3-A4A9-C21A32AACBEC} = {BB931C9A-EE48-42A6-B5F3-3A193B1A051A}
 		{138E97ED-8C5B-4C31-90F8-4260714A56E7} = {BB931C9A-EE48-42A6-B5F3-3A193B1A051A}
+		{C6E4069B-2305-4CDA-9F4E-0B79A8599935} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
+		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6E8897DC-99C6-4585-A2BA-CB194F3B1AB7}
