+++ src/Miners/CCMinerMTP/CCMinerMTPPlugin.cs
@@ -1,0 +1,71 @@
+﻿using MinerPluginToolkitV1;
+using MinerPluginToolkitV1.Interfaces;
+using NHM.Common.Algorithm;
+using NHM.Common.Device;
+using NHM.Common.Enums;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace CCMinerMTP
+{
+    public abstract class CCMinerMTPPlugin : PluginBase
+    {
+        public CCMinerMTPPlugin()
+        {
+            // set default internal settings
+            MinerOptionsPackage = PluginInternalSettings.MinerOptionsPackage;
+            DefaultTimeout = PluginInternalSettings.DefaultTimeout;
+            GetApiMaxTimeoutConfig = PluginInternalSettings.GetApiMaxTimeoutConfig;
+        }
+
+        //public override string PluginUUID => "MISSING";
+
+        public override Version Version => new Version(2, 0);
+        public override string Name => "CCMinerMTP";
+
+        public override string Author => "stanko@nicehash.com";
+
+        public override Dictionary<BaseDevice, IReadOnlyList<Algorithm>> GetSupportedAlgorithms(IEnumerable<BaseDevice> devices)
+        {
+            var supported = new Dictionary<BaseDevice, IReadOnlyList<Algorithm>>();
+            var reqCudaVer = Checkers.CudaVersion.CUDA_10_0_130;
+            var isCompatible = Checkers.IsCudaCompatibleDriver(reqCudaVer, CUDADevice.INSTALLED_NVIDIA_DRIVERS);
+            if (!isCompatible) return supported; // return emtpy
+
+            var cudaGpus = devices
+                .Where(dev => dev is CUDADevice gpu && gpu.SM_major >= 5)
+                .Cast<CUDADevice>();
+
+            foreach (var gpu in cudaGpus)
+            {
+                var algorithms = new List<Algorithm> {
+                    new Algorithm(PluginUUID, AlgorithmType.MTP) { Enabled = false }
+                };
+                var filteredAlgorithms = Filters.FilterInsufficientRamAlgorithmsList(gpu.GpuRam, algorithms);
+                if (filteredAlgorithms.Count > 0) supported.Add(gpu, filteredAlgorithms);
+            }
+
+            return supported;
+        }
+
+        protected override MinerBase CreateMinerBase()
+        {
+            return new CCMinerMTP(PluginUUID);
+        }
+
+        public override IEnumerable<string> CheckBinaryPackageMissingFiles()
+        {
+            var miner = CreateMiner() as IBinAndCwdPathsGettter;
+            if (miner == null) return Enumerable.Empty<string>();
+            var pluginRootBinsPath = miner.GetBinAndCwdPaths().Item2;
+            return BinaryPackageMissingFilesCheckerHelpers.ReturnMissingFiles(pluginRootBinsPath, new List<string> { "ccminer.exe" });
+        }
+
+        public override bool ShouldReBenchmarkAlgorithmOnDevice(BaseDevice device, Version benchmarkedPluginVersion, params AlgorithmType[] ids)
+        {
+            //no new version available
+            return false;
+        }
+    }
+}
