+++ src/Miners/CCMinerTpruvot/CCMinerTpruvot.cs
@@ -1,21 +1,28 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using MinerPlugin;
 using NHM.Common;
 using NHM.Common.Enums;
+using MinerPlugin;
+using MinerPluginToolkitV1.CCMinerCommon;
 
-namespace NiceHashMiner.Mining.Plugins
+namespace CCMinerTpruvot
 {
-    class CCMinerIntegratedMiner : MinerPluginToolkitV1.CCMinerCommon.CCMinerBase
+    public class CCMinerTpruvot : CCMinerBase
     {
-        public CCMinerIntegratedMiner(string uuid, string dirPath) : base(uuid)
+        public CCMinerTpruvot(string uuid) : base(uuid)
+        { }
+
+        protected override string AlgorithmName(AlgorithmType algorithmType)
         {
-            _noTimeLimitOption = "ccminer_klaust" == dirPath;
+            switch (algorithmType)
+            {
+                case AlgorithmType.X16R: return "x16r";
+                case AlgorithmType.Lyra2REv3: return "lyra2v3";
+            }
+            // TODO throw exception
+            return "";
         }
 
         public override async Task<BenchmarkResult> StartBenchmark(CancellationToken stop, BenchmarkPerformanceType benchmarkType = BenchmarkPerformanceType.Standard)
@@ -37,27 +44,8 @@ namespace NiceHashMiner.Mining.Plugins
             return ret;
         }
 
-#pragma warning disable 0618
-        protected override string AlgorithmName(AlgorithmType algorithmType)
-        {
-            switch (algorithmType)
-            {
-                case AlgorithmType.X16R: return "x16r";
-                case AlgorithmType.Lyra2REv3: return "lyra2v3";
-                case AlgorithmType.MTP: return "mtp";
-            }
-            // TODO throw exception
-            return "";
-        }
-#pragma warning restore 0618
-
         public override Tuple<string, string> GetBinAndCwdPaths()
         {
-            if (_uuid != "CCMinerTpruvot")
-            {
-                return base.GetBinAndCwdPaths();
-            }
-
             var pluginRootBins = Paths.MinerPluginsPath(_uuid, "bins");
             var binPath = Path.Combine(pluginRootBins, "ccminer-x64.exe");
             return Tuple.Create(binPath, pluginRootBins);
