+++ src/Miners/CCMinerTpruvot/PluginInternalSettings.cs
@@ -1,0 +1,10 @@
+﻿using MinerPluginToolkitV1.CCMinerCommon;
+using MinerPluginToolkitV1.ExtraLaunchParameters;
+
+namespace CCMinerTpruvot
+{
+    internal static class PluginInternalSettings
+    {
+        internal static MinerOptionsPackage MinerOptionsPackage = CCMinerOptionsPackage.MinerOptionsPackage;
+    }
+}
