+++ src/Miners/MinerPluginToolkitV1/CCMinerCommon/CCMinerAPIHelpers.cs
@@ -4,9 +4,6 @@ using NHM.Common.Enums;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Net.Sockets;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace MinerPluginToolkitV1.CCMinerCommon
