+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.Mining.cs
@@ -10,6 +10,7 @@ using NiceHashMiner.Mining;
 using NiceHashMiner.Stats;
 using NHM.Common;
 using NiceHashMiner.Configs;
+using NiceHashMiner.Utils;
 
 namespace NiceHashMiner
 {
