--- src/NiceHashMiner/Mining/Plugins/IntegratedPlugins/CCMinerIntegrated/CCMinerMTPIntegratedPlugin.cs
@@ -1,56 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using MinerPlugin;
-using MinerPluginToolkitV1;
-using MinerPluginToolkitV1.Configs;
-using MinerPluginToolkitV1.Interfaces;
-using NHM.Common.Algorithm;
-using NHM.Common.Device;
-using NHM.Common.Enums;
-
-namespace NiceHashMiner.Mining.Plugins
-{
-    class CCMinerMTPIntegratedPlugin : CCMinersPluginBase
-    {
-        public override string PluginUUID => "CCMinerMTP";
-
-        public override Version Version => new Version(1, 1);
-
-        public override string Name => "CCMinerMTP";
-
-        protected override string DirPath => "ccminer_mtp";
-
-        public override Dictionary<BaseDevice, IReadOnlyList<Algorithm>> GetSupportedAlgorithms(IEnumerable<BaseDevice> devices)
-        {
-            var supported = new Dictionary<BaseDevice, IReadOnlyList<Algorithm>>();
-            var reqCudaVer = Checkers.CudaVersion.CUDA_10_0_130;
-            var isCompatible = Checkers.IsCudaCompatibleDriver(reqCudaVer, CUDADevice.INSTALLED_NVIDIA_DRIVERS);
-            if (!isCompatible) return supported; // return emtpy
-
-            var cudaGpus = devices
-                .Where(dev => dev is CUDADevice gpu && gpu.SM_major >= 5)
-                .Cast<CUDADevice>();
-
-            foreach (var gpu in cudaGpus)
-            {
-                var algorithms = new List<Algorithm> {
-                    new Algorithm(PluginUUID, AlgorithmType.MTP) { Enabled = false }
-                };
-                supported.Add(gpu, algorithms);
-            }
-
-            return supported;
-        }
-
-        #region IGetApiMaxTimeoutV2
-
-        public new TimeSpan GetApiMaxTimeout(IEnumerable<MiningPair> miningPairs)
-        {
-            return MinerApiMaxTimeoutSetting.ParseMaxTimeout(new TimeSpan(0,2,0), _getApiMaxTimeoutConfig, miningPairs);
-        }
-        #endregion IGetApiMaxTimeoutV2
-    }
-}
