+++ src/NiceHashMiner/Stats/NiceHashStats_production_old.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using WebSocketSharp;
 using NiceHashMiner.Stats.Models;
+using NiceHashMiner.Utils;
 
 namespace NiceHashMiner.Stats
 {
